Updates packages
