+++ back/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "back",
+  "version": "1.0.0",
+  "description": "README back",
+  "main": "mongoDbClient.js",
+  "scripts": {
+    "test": "node index.test"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "mongodb": "^4.1.0",
+    "tape": "^5.3.1"
+  }
+}
